ticket: #28  refactor httpMethod constants to avoid naming conflict
because this project uses a lot of common terms we are avoid having domain names that could clash with the naming of other libraries. it is easy to have UaiHttpCode and knowing that this is a common domain of the project than using HttpCode and clash with several other libs.