ticket: #28 - create a webServer simulator
this is a helper service to receive a rest request. with this class we could start a fast/small webServer during our unit tests to validate if we received or not a http request. this will help for now on the callback requests that we should validate if we received it or not